Switch tests to Typescript and add extractor
extractor is 'extension-code' free, so our test harness still works.
Otherwise polyfill would have been run in non-extension context and
that's apparently not allowed.
