+++ package.json
@@ -5,8 +5,12 @@
   "devDependencies": {
     "@babel/core": "^7.17.0",
     "@babel/preset-env": "^7.16.11",
+    "@rollup/plugin-commonjs": "^21.0.2",
     "@rollup/plugin-node-resolve": "^13.1.3",
     "@rollup/plugin-typescript": "^8.3.1",
+    "@types/jest": "^27.4.1",
+    "@types/jest-environment-puppeteer": "^5.0.0",
+    "@types/puppeteer": "^5.4.5",
     "@types/webextension-polyfill": "^0.8.2",
     "@typescript-eslint/eslint-plugin": "^5.14.0",
     "@typescript-eslint/parser": "^5.14.0",
@@ -19,8 +23,10 @@
     "jest": "^27.5.0",
     "jest-puppeteer": "^6.1.0",
     "node-fetch": "^3.2.0",
+    "puppeteer": "^13.5.1",
     "puppeteer-core": "^13.1.3",
     "rollup": "^2.42.0",
+    "ts-jest": "^27.1.3",
     "typescript": "^4.6.2",
     "webextension-polyfill": "^0.8.0"
   },
@@ -28,7 +34,7 @@
     "start": "rollup -c -w",
     "build": "rollup -c",
     "lint": "eslint",
-    "test": "jest",
+    "test": "jest --runInBand",
     "version": "gsed -i '4s/\"version\": \".*\",/\"version\": \"'\"$(jq -r .version package.json)\"'\",/' manifest.json && git add manifest.json"
   },
   "repository": {
