+++ rollup.config.js
@@ -1,4 +1,5 @@
 import typescript from '@rollup/plugin-typescript'
+import commonjs from '@rollup/plugin-commonjs'
 import { nodeResolve } from '@rollup/plugin-node-resolve'
 
 export default [
@@ -8,7 +9,7 @@ export default [
       file: 'build/background.js',
       format: 'cjs'
     },
-    plugins: [nodeResolve(), typescript()]
+    plugins: [commonjs(), nodeResolve(), typescript()]
   },
   {
     input: 'src/content.ts',
@@ -16,7 +17,7 @@ export default [
       file: 'build/content.js',
       format: 'cjs'
     },
-    plugins: [nodeResolve(), typescript()]
+    plugins: [commonjs(), nodeResolve(), typescript()]
   },
   {
     input: 'src/options.ts',
@@ -24,7 +25,7 @@ export default [
       file: 'build/options.js',
       format: 'cjs'
     },
-    plugins: [nodeResolve(), typescript()]
+    plugins: [commonjs(), nodeResolve(), typescript()]
   },
   {
     input: 'src/popup.ts',
@@ -32,13 +33,22 @@ export default [
       file: 'build/popup.js',
       format: 'cjs'
     },
-    plugins: [nodeResolve(), typescript()]
+    plugins: [commonjs(), nodeResolve(), typescript()]
+  },
+  {
+    input: 'tests/content_test.ts',
+    output: {
+      file: 'test_build/content_test.js',
+      format: 'cjs'
+    },
+    plugins: [commonjs(), nodeResolve(), typescript()]
+  },
+  {
+    input: 'tests/sites.test.ts',
+    output: {
+      file: 'test_build/sites.test.js',
+      format: 'cjs'
+    },
+    plugins: [commonjs(), nodeResolve(), typescript()]
   }
-  // {
-  //   input: 'tests/content_test.js',
-  //   output: {
-  //     file: 'test_build/content_test.js',
-  //     format: 'cjs'
-  //   }
-  // }
 ]
