+++ tests/index.d.ts
@@ -1,0 +1,7 @@
+import { ExtractorInterface } from '../src/types.js'
+
+declare global {
+    interface Window {
+        extractor: ExtractorInterface
+    }
+}
