+++ tsconfig.json
@@ -4,5 +4,5 @@
     "allowJs": true,
     "target": "es6"
   },
-  "include": ["./src/**/*"]
+  "include": ["./src/**/*", "./tests/**/*"]
 }
